Multiple commit
- Changed io.h function return types to int so errors can be returned with integer values
- Moved tutorials into a tutorials folder
- Altered variable types in waveform struct
- Began io.c implementation, now checks for file existance and opens

+++ PowerQualityWaveformAnalyser/include/waveform.h
@@ -8,13 +8,13 @@
 
 typedef struct{
     float timestamp;
-    float phaseAVoltage;
-    float phaseBVoltage;
-    float phaseCVoltage;
-    float lineCurrent;
-    float frequency;
-    float powerFactor;
-    float thdPercent;
+    double phaseAVoltage;
+    double phaseBVoltage;
+    double phaseCVoltage;
+    double lineCurrent;
+    double frequency;
+    double powerFactor;
+    double thdPercent;
 } waveform;
 
 // Function prototypes for waveform calculations
